+++ arpwatch_unix/null.go
@@ -1,0 +1,9 @@
+// +build windows
+
+// for avoiding build errors on windows
+package arpwatch_unix
+
+var RlogServer string
+
+func LogToRemote(message string, rServer string) {
+}
